requirements.txt
pandas>=1.5.0
kafka-python>=2.0.2
mysql-connector-python>=8.0.0
pymysql>=1.0.0
sqlalchemy>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
